fix: G2Jacobian::add() missing factor of 2 in r computation
The add-2007-bl formula requires r = 2*(S2-S1) but the code had
r = S2-S1 (missing the doubling). This caused incorrect results in:
- G2 MSM bucket reduction (Jacobian+Jacobian additions)
- G2 MSM window combination
- G2Jacobian::scalar_mul

The mixed addition (add_affine) was already correct.
Found by code review agent.

+++ sp1-gpu/crates/groth16/src/g2.rs
@@ -122,7 +122,7 @@ impl G2Jacobian {
         let s2 = other.y * z1z1 * self.z;
 
         let h = u2 - u1;
-        let r = s2 - s1;
+        let r = (s2 - s1).double(); // add-2007-bl requires r = 2*(S2-S1)
 
         if h.is_zero() {
             if r.is_zero() {
